feat: move structs to a separate file for better organization
